+++ contracts/aggregator/src/test/events.rs
@@ -1,8 +1,5 @@
 use soroban_sdk::{testutils::{Events}, vec, IntoVal, symbol_short, String}; 
-use soroban_sdk::{Address, testutils::Address as _};
-use crate::error::AggregatorError;
-use crate::test::{SoroswapAggregatorTest, create_protocols_addresses};
-use crate::test::get_proxies::new_update_proxies_addresses;
+use crate::test::{SoroswapAggregatorTest, create_protocols_addresses, new_update_proxies_addresses};
 
 use crate::event::{
     InitializedEvent,
